Issue #1 fixed. Project ready 100%

+++ Week3_LocalStorage_Project/Resources/CRUD/QueryDELETE.js
@@ -49,22 +49,63 @@ function deleteData(deleteAll, idTarget){
     }else if (deleteAll == false){
         
         // Message for users that tell them that the selected rows had been deleted...
-        var messageDialog2 = Ti.UI.createAlertDialog({
-            message: "Your selected password was deleted...",
-            ok: "Dismiss",
-            title: "Delete operation completed"
-        }).show();
+        var messageDialog2 = Ti.UI.createOptionDialog({
+            cancel: 1,
+            options: ["Delete", "Cancel"],
+            selectedIndex: 1,
+            destructive: 0,
+            title: "Are you sure you want to delete the selected password???"
+        });
         
-        datab.execute("DELETE FROM accounts WHERE rowid=" + idTarget);
+        // Setting the event listener to the option dialog
+        messageDialog2.addEventListener("click", function(e){
+            
+            // Conditional that verifies what option was selected from the dialog button 
+            if ( e.index == e.source.cancel){
+                
+                 var messageDialog3 = Ti.UI.createAlertDialog({
+                         message: "Your selected password was NOT deleted...",
+                         ok: "Dismiss",
+                         title: "Delete operation canceled"
+                }).show();
+                
+                // Calling the methods and functions to: close the database's rows, close the entire database and refresh the app
+                TableRows.close();
+                datab.close();
+                QController.initController(null, 0, 1, 1, 62.5, 0);
+            
+            }else{
+            
+                var messageDialog4 = Ti.UI.createAlertDialog({
+                         message: "Your selected password was deleted...",
+                         ok: "Dismiss",
+                         title: "Delete operation completed"
+                }).show();
+                
+                // Executing the delete c
+                datab.execute("DELETE FROM accounts WHERE rowid=" + idTarget);
+                
+                // Calling the methods and functions to: console log, close the database's rows, close the entire database and refresh the app   
+                displayTableOnConsole();
+                TableRows.close();
+                datab.close();
+                QController.initController(null, 0, 1, 1, 62.5, 0);
+            };
         
-        displayTableOnConsole();
+        });
         
-        TableRows.close();
-        datab.close();
+        // Calling an option dialog to show when the user tries to delete an item... 
+        messageDialog2.show();
     
     }else{
         
         // Desabled feature and work is in process. Bugs and unidetified scopes had not been found yet...
+        
+        // IMPORTANT NOTE:
+        // Change Log: Friday February 21, 2014 at 20:32 EST
+        // The following feature is no longer necesary for this application... Att GchrisWill
+        
+        
         var DialogOptions = {
             cancel: 1,
             options: ["Delete", "Cancel"],
@@ -73,17 +114,27 @@ function deleteData(deleteAll, idTarget){
             title: "Are you sure you want to DELETE all your saved passwords???"
         };
         
-        var warningDialog = Ti.UI.createOptionDialog(DialogOptions).show();
+        var warningDialog = Ti.UI.createOptionDialog(DialogOptions);
         
+        // ISSUE NOTE!
         // Issue #1: Event Listener is entering in an undefined scope that has not been identified yet...
         // Detail: When the event is trigered, the "delete all" feature breaks outputing and error to the console stating that the "warnign dialog" is undefined
         // and that is not an object..
         // The "delete" all feature has been censured until the bug is found and fixed...
+        
+        // IMPORTANT NOTE:
+        // Change Log: Friday February 21, 2014 at 20:14 EST
+        // Bug has been found and we had to spray it with "Blakc Jack" Insect Killer. The bug has been successfuly exterminated and application is fully up and running 100%...
+        
+        // Issue was that the createOptionDialog was set with the show() method and the event listener was unable to identify it as an object
+        // due to the show() method call before setting the event listener to the object. Now the show() method call has been set after the setting the event listener...
+        
+        // Issue #1 CLOSE!
         warningDialog.addEventListener("click", function(e){
             
             if(e.index == e.source.destructive){
                 
-                var messageDialog3 = Ti.UI.createAlertDialog({
+                var messageDialog5 = Ti.UI.createAlertDialog({
                      message: "All of your passwords were deleted...",
                      ok: "Dismiss",
                      title: "Delete operation completed"
@@ -95,7 +146,7 @@ function deleteData(deleteAll, idTarget){
                 
             }else{
                  
-                 var messageDialog4 = Ti.UI.createAlertDialog({
+                 var messageDialog6 = Ti.UI.createAlertDialog({
                      message: "Your passwords weren't deleted...",
                      ok: "Dismiss",
                      title: "Delete operation canceled"
@@ -107,7 +158,10 @@ function deleteData(deleteAll, idTarget){
             datab.close();
         
         });
+        
+        warningDialog.show();
     };
+    
 };
 
 exports.databaseDELETE = deleteData;
